refactor: self contain user summary and move modals
