+++ src/users/UserSummary.test.jsx
@@ -96,142 +96,4 @@ describe('User Summary Component Tests', () => {
       expect(rowValue.text()).toEqual('N/A');
     });
   });
-
-  describe('Disable User Button', () => {
-    it('Disable User button for active user', () => {
-      const passwordActionButton = wrapper.find('#toggle-password').hostNodes();
-      expect(passwordActionButton.text()).toEqual('Disable User');
-      expect(passwordActionButton.disabled).toBeFalsy();
-    });
-
-    it('Disable User Modal', () => {
-      const mockApiCall = jest.spyOn(api, 'postTogglePasswordStatus').mockImplementation(() => {});
-      const passwordActionButton = wrapper.find('#toggle-password').hostNodes();
-      let disableDialogModal = wrapper.find('Modal#user-account-status-toggle');
-
-      expect(disableDialogModal.prop('open')).toEqual(false);
-      expect(passwordActionButton.text()).toEqual('Disable User');
-      expect(passwordActionButton.disabled).toBeFalsy();
-
-      passwordActionButton.simulate('click');
-      disableDialogModal = wrapper.find('Modal#user-account-status-toggle');
-
-      expect(disableDialogModal.prop('open')).toEqual(true);
-      expect(disableDialogModal.prop('title')).toEqual('Disable user confirmation');
-      disableDialogModal.find('input[name="comment"]').simulate('change', { target: { value: 'Disable Test User' } });
-      disableDialogModal.find('button.btn-danger').hostNodes().simulate('click');
-
-      expect(UserSummaryData.changeHandler).toHaveBeenCalled();
-      disableDialogModal.find('button.btn-link').simulate('click');
-      disableDialogModal = wrapper.find('Modal#user-account-status-toggle');
-      expect(disableDialogModal.prop('open')).toEqual(false);
-      mockApiCall.mockRestore();
-    });
-  });
-
-  describe('Enable User Button', () => {
-    beforeEach(() => {
-      const passwordStatusData = { ...UserSummaryData.userData.passwordStatus, status: 'Unusable' };
-      const userData = { ...UserSummaryData.userData, passwordStatus: passwordStatusData };
-      mountUserSummaryWrapper({ ...UserSummaryData, userData });
-    });
-
-    it('Enable User button for disabled user', () => {
-      const passwordActionButton = wrapper.find('#toggle-password').hostNodes();
-      expect(passwordActionButton.text()).toEqual('Enable User');
-      expect(passwordActionButton.disabled).toBeFalsy();
-    });
-
-    it('Enable User Modal', () => {
-      const mockApiCall = jest.spyOn(api, 'postTogglePasswordStatus').mockImplementation(() => {});
-      const passwordActionButton = wrapper.find('#toggle-password').hostNodes();
-      let enableUserModal = wrapper.find('Modal#user-account-status-toggle');
-
-      expect(enableUserModal.prop('open')).toEqual(false);
-      expect(passwordActionButton.text()).toEqual('Enable User');
-      expect(passwordActionButton.disabled).toBeFalsy();
-
-      passwordActionButton.simulate('click');
-      enableUserModal = wrapper.find('Modal#user-account-status-toggle');
-
-      expect(enableUserModal.prop('open')).toEqual(true);
-      expect(enableUserModal.prop('title')).toEqual('Enable user confirmation');
-      enableUserModal.find('input[name="comment"]').simulate('change', { target: { value: 'Enable Test User' } });
-      enableUserModal.find('button.btn-danger').hostNodes().simulate('click');
-
-      expect(UserSummaryData.changeHandler).toHaveBeenCalled();
-      mockApiCall.mockRestore();
-    });
-  });
-
-  describe('Reset Password Button', () => {
-    it('Reset Password button for a User', () => {
-      const passwordResetButton = wrapper.find('#reset-password').hostNodes();
-      expect(passwordResetButton.text()).toEqual('Reset Password');
-    });
-
-    it('Reset Password Modal', () => {
-      const mockApiCall = jest.spyOn(api, 'postResetPassword').mockImplementation(() => {});
-      const passwordResetButton = wrapper.find('#reset-password').hostNodes();
-      let resetPasswordModal = wrapper.find('Modal#user-account-reset-password');
-
-      expect(resetPasswordModal.prop('open')).toEqual(false);
-      expect(passwordResetButton.text()).toEqual('Reset Password');
-
-      passwordResetButton.simulate('click');
-      resetPasswordModal = wrapper.find('Modal#user-account-reset-password');
-
-      expect(resetPasswordModal.prop('open')).toEqual(true);
-      expect(resetPasswordModal.prop('title')).toEqual('Reset Password');
-      const confirmLabel = resetPasswordModal.find('label');
-      expect(confirmLabel.text()).toContain('Do you wish to proceed?');
-      resetPasswordModal.find('button.btn-danger').hostNodes().simulate('click');
-
-      expect(UserSummaryData.changeHandler).toHaveBeenCalled();
-      resetPasswordModal.find('button.btn-link').simulate('click');
-      resetPasswordModal = wrapper.find('Modal#user-account-reset-password');
-      expect(resetPasswordModal.prop('open')).toEqual(false);
-      mockApiCall.mockRestore();
-    });
-  });
-
-  describe('Password Toggle History', () => {
-    beforeEach(() => {
-      const passwordHistory = [
-        {
-          created: Date().toLocaleString(),
-          comment: 'Test Disabled',
-          disabled: false,
-          createdBy: 'staff',
-        },
-        {
-          created: Date().toLocaleString(),
-          comment: 'Test Enable',
-          disabled: true,
-          createdBy: 'staff',
-        },
-      ];
-      const passwordStatusData = { ...UserSummaryData.userData.passwordStatus, passwordToggleHistory: passwordHistory };
-      const userData = { ...UserSummaryData.userData, passwordStatus: passwordStatusData };
-      mountUserSummaryWrapper({ ...UserSummaryData, userData });
-    });
-    it('Password History Modal', () => {
-      const passwordHistoryButton = wrapper.find('button#toggle-password-history');
-      let historyModal = wrapper.find('Modal#password-history');
-
-      expect(historyModal.prop('open')).toEqual(false);
-      expect(passwordHistoryButton.text()).toEqual('Show History');
-      expect(passwordHistoryButton.disabled).toBeFalsy();
-
-      passwordHistoryButton.simulate('click');
-      historyModal = wrapper.find('Modal#password-history');
-
-      expect(historyModal.prop('open')).toEqual(true);
-      expect(historyModal.find('table tbody tr')).toHaveLength(2);
-
-      historyModal.find('button.btn-link').simulate('click');
-      historyModal = wrapper.find('Modal#password-history');
-      expect(historyModal.prop('open')).toEqual(false);
-    });
-  });
 });
